requirements.txt
asgiref==3.6.0
# backports.zoneinfo==0.2.1
Django==4.1.4
django-admin-interface==0.24.2
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.8.1
django-filter==22.1
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.3.0
python-slugify==7.0.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7
whitenoise==6.2.0
